feat(cli): Add functionality to check diff for entry on current PR (#96)
* add wip for check-diff functionality

* add changelog entry

* complete functionality

* imp(cli): Refactor Git operations into dedicated module (#97)

* refactor git from github module

* add changelog entry

* address TODOs and use u64 for pr number to align with octocrab

* add some logs for failed diff check

* move git related logic from diff checking to the new git submodule

+++ src/inputs.rs
@@ -21,11 +21,11 @@ pub fn get_change_type(config: &Config, suggestion: &str) -> Result<String, Inpu
     )
 }
 
-pub fn get_pr_number(default_value: u16) -> Result<u16, InputError> {
+pub fn get_pr_number(default_value: u64) -> Result<u64, InputError> {
     Ok(Text::new("Please provide the PR number:")
         .with_initial_value(format!("{}", &default_value).as_str())
         .prompt()?
-        .parse::<u16>()?)
+        .parse::<u64>()?)
 }
 
 pub fn get_category(config: &Config, suggestion: &str) -> Result<String, InputError> {
